+++ src/main/java/site/siredvin/gttruesteam/TrueSteamRecipeTypes.java
@@ -26,6 +26,7 @@ import site.siredvin.gttruesteam.machines.coating_shrine.CoatingShrine;
 import site.siredvin.gttruesteam.machines.cooling_box.CoolingBox;
 import site.siredvin.gttruesteam.machines.industrial_gas_pressurizer.IndustrialGasPressurizer;
 import site.siredvin.gttruesteam.machines.industrial_heater.InfernalBoiler;
+import site.siredvin.gttruesteam.machines.spawner_extraction.SpawnerExtractionMachine;
 import site.siredvin.gttruesteam.recipe.condition.CoatingFluidCondition;
 
 import java.util.ArrayList;
@@ -40,6 +41,7 @@ public class TrueSteamRecipeTypes {
     public static String OVERHEATED_KEY = "overheated";
     public static String COOLING_CONSUMED = "cooling_consumed";
     public static String ASSUMED_PERCENTAGE = "assumed_percentage";
+    public static String LOOT_TABLE_DROPS = "loot_table_drops";
 
     public static String OVERHEATED_WIDGET_ID = "overheatedWidget";
     public static String INTERNAL_CYCLED_WIDGET_ID = "infernalCyclesWidget";
@@ -136,6 +138,13 @@ public class TrueSteamRecipeTypes {
                     yOffset = 20 * (i / 3);
                 }
             });
+
+    public static GTRecipeType SPAWNER_EXTRACTION = register("spawner_extraction", GTRecipeTypes.MULTIBLOCK,
+            "Spawner extraction")
+            .setMaxIOSize(3, 2, 2, 2)
+            .setEUIO(IO.IN)
+            .setProgressBar(GuiTextures.PROGRESS_BAR_ARROW, LEFT_TO_RIGHT)
+            .setIconSupplier(() -> SpawnerExtractionMachine.MACHINE.getItem().getDefaultInstance());
 
     @SuppressWarnings("deprecation")
     public static GTRecipeType register(String name, String group, String displayName, RecipeType<?>... proxyRecipes) {